conformance/buffimgs/cl_commandQueue_enqueueReadBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueReadBuffer conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue :: enqueueReadBuffer() test case.");
    debug("( WebCLBuffer buffer, CLboolean blockingRead, CLuint bufferOffset, CLuint numBytes, ArrayBuffer hostPtr, optional WebCLEvent[]? eventWaitList, optional WebCLEvent? event)");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var arrBuffer = new Float32Array(BUFFER_SIZE);
    var numberBytes = DATA_SIZE;
    var bufferOffset = 0;

    // MEM_READ_ONLY Buffers.
    var webCLBufferRead = webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);
    shouldBeTrue("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferRead, true, bufferOffset, numberBytes, arrBuffer)");
    shouldBeTrue("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferRead, false, bufferOffset, numberBytes, arrBuffer)");

    // MEM_WRITE_ONLY Buffers.
    var webCLBufferWrite = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferWrite, true, bufferOffset, numberBytes, arrBuffer)", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferWrite, false, bufferOffset, numberBytes, arrBuffer)", "INVALID_VALUE");

    // MEM_READ_WRITE Buffers.
    var webCLBufferReadWrite = webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE);
    shouldBeTrue("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffset, numberBytes, arrBuffer)");
    shouldBeTrue("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, false, bufferOffset, numberBytes, arrBuffer)");

    // FIXME :: Need to add test cases passing event parameters.

    // Testing invalid objects as params.
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, invalid_object, true, bufferOffset, numberBytes, arrBuffer);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, invalid_object, bufferOffset, numberBytes, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, invalid_object, numberBytes, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffset, invalid_object, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffset, numberBytes, invalid_object);", "INVALID_VALUE");

    // Testing for invalid value of bufferOffset.
    var bufferOffsetInvalid = -1;
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffsetInvalid, numberBytes, arrBuffer);", "INVALID_VALUE");

    // Testing for invalid value of numBytes.
    var numberBytesInvalid = DATA_SIZE + 1;
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffset, numberBytesInvalid, arrBuffer);", "INVALID_VALUE");

    // Testing for invalid value of host Pointer.
    var arrBufferInvalid = 10;
    shouldThrowExceptionName("wtu.enqueueReadBuffer(webCLCommandQueue, webCLBufferReadWrite, true, bufferOffset, numberBytes, arrBufferInvalid);", "INVALID_VALUE");

} catch(e) {
    testFailed("WebCLCommandQueue :: enqueueReadBuffer() throws an exception:" + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
